% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{aggregate_samples}
\alias{aggregate_samples}
\title{Aggregate expression data from all samples}
\usage{
aggregate_samples(obj, max_cells = 12000)
}
\arguments{
\item{obj}{'phemdObj' object containing raw expression data and associated metadata}

\item{max_cells}{Maximum number of cells across all samples to be included in final matrix on which Monocle 2 will be run}
}
\value{
Same as input 'phemdObj' object with additional slot 'data_aggregate' containing aggregated expression data (num_markers x num_cells)
}
\description{
Takes initial phemdObj and returns phemdObj with additional data frame in slot @data_aggregate containing cells aggregated from all samples (to be used for further analyses e.g. Monocle 2 trajectory building / pseudotime mapping / cell clustering)
}
\details{
Subsamples cells as necessary based on \code{max_cells}. If subsampling is performed, an equal number of cells are subsampled from each sample
}
\examples{
\dontrun{
my_phemdObj <- create_dataobj(all_expn_data, genes_measured, my_sample_names)
my_phemdObj_lg <- remove_tiny_samples(my_phemdObj, 10) #removes samples with fewer than 10 cells
my_phemdObj_lg <- aggregate_samples(my_phemdObj_lg, max_cells=10000)
}
}
