% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-core.R
\name{aggregateSamples}
\alias{aggregateSamples}
\title{Aggregate expression data from all samples}
\usage{
aggregateSamples(obj, max_cells = 12000)
}
\arguments{
\item{obj}{'Phemd' object containing raw expression data and associated metadata}

\item{max_cells}{Maximum number of cells across all samples to be included in final matrix on which Monocle 2 will be run}
}
\value{
Same as input 'Phemd' object with additional slot 'data_aggregate' containing aggregated expression data (num_markers x num_cells)
}
\description{
Takes initial Phemd object and returns object with additional data frame in slot @data_aggregate containing cells aggregated from all samples (to be used for further analyses e.g. Monocle 2 trajectory building / pseudotime mapping / cell clustering)
}
\details{
Subsamples cells as necessary based on \code{max_cells}. If subsampling is performed, an equal number of cells are subsampled from each sample
}
\examples{
my_phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames))
my_phemdObj_lg <- removeTinySamples(my_phemdObj, 10)
my_phemdObj_lg <- aggregateSamples(my_phemdObj_lg, max_cells=1000)

}
