\name{all_genes}
\alias{all_genes}
\docType{data}
\title{All genes included in (subsampled) melanoma single-cell RNA-seq expression data}
\description{This object contains 100 genes measured in melanoma single-cell RNA-seq expression data.}
\usage{
  data(all_genes)
}
\format{
  Vector of length 100 representing row names of each matrix in melanoma expression dataset
}
\source{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE72056}
\references{Tirosh, I. et al. Dissecting the multicellular ecosystem of metastatic melanoma by single-cell RNA-seq. Science 352, 189–196 (2016)}
\keyword{datasets}
