% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-core.R
\name{assignCellClusterNearestNode}
\alias{assignCellClusterNearestNode}
\title{Assign cells to a reference cell subtype}
\usage{
assignCellClusterNearestNode(cur_cells, ref_cells, ref_cell_labels,
  cell_model = c("monocle2", "seurat", "phate"))
}
\arguments{
\item{cur_cells}{Matrix of cells to be assigned to clusters (Dim: \var{num_cells} x \var{num_markers})}

\item{ref_cells}{Matrix of cells used to build reference Monocle 2 tree (Dim: \var{num_monocle_cells} x \var{num_markers})}

\item{ref_cell_labels}{Vector of length \var{num_monocle_cells} containing Monocle 2 cell branch assignments}

\item{cell_model}{Either "monocle2", "seurat", or "phate" depending on method used to model cell state space}
}
\value{
Vector of length \var{num_cells} representing cluster assignments for each cell in \var{cur_cells}
}
\description{
Assigns each cell in \code{cur_cells} to a cluster based on nearest cell in Monocle 2 tree
}
\details{
Private method (not exported in namespace). Uses RANN package for fast knn search
}
\examples{
\dontrun{
cur_cells_cluster_labels <- assignCellClusterNearestNode(cur_cells_expn_data,
clustered_cells_expn_data, clustered_cells_cluster_labels, cell_model='monocle2')
}
}
