% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{attach_seuratobj}
\alias{attach_seuratobj}
\title{Attach 'seurat' object to 'phemdObj' object}
\usage{
attach_seuratobj(phemd_obj, seurat_obj, batch.colname = "plt")
}
\arguments{
\item{phemd_obj}{phemdObj struct initialized using create_dataobj}

\item{seurat_obj}{S4 'seurat' object containing batch-normalized reference cell data}

\item{batch.colname}{Name of column in Seurat object that denotes batch ID}
}
\value{
'phemdObj' object containing with attached Seurat object
}
\description{
Allows user to attach batch-normalized reference cell data from Seurat into 'phemdObj' object containing raw expression data and metadata
}
\examples{
\dontrun{
my_phemdObj <- attach_seuratobj(my_phemdObj, my_seuratObj)
}
}
