% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phemd-methods.R
\name{batchIDs}
\alias{batchIDs}
\title{Accessor function for batch ID for each sample}
\usage{
batchIDs(obj)
}
\arguments{
\item{obj}{Phemd object}
}
\value{
Vector of length num_samples representing the experiment (batch) in which the sample was profiled
}
\description{
Accessor function for batch ID for each sample
}
\examples{
phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames))
batch_metadata <- batchIDs(phemdObj)

}
