% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phemd-methods.R
\name{celltypeFreqs}
\alias{celltypeFreqs}
\title{Accessor function for cell subtype distribution for each sample}
\usage{
celltypeFreqs(obj)
}
\arguments{
\item{obj}{Phemd object}
}
\value{
Matrix representing cell subtype relative frequencies for each sample (num_samples x num_genes)
}
\description{
Accessor function for cell subtype distribution for each sample
}
\examples{
phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames_data))
celltype_weights <- celltypeFreqs(phemdObj)

}
