% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cluster_individual_samples}
\alias{cluster_individual_samples}
\title{Computes cell subtype abundances for each sample}
\usage{
cluster_individual_samples(obj, verbose = FALSE,
  cell_model = "monocle2")
}
\arguments{
\item{obj}{'phemdObj' object containing single-cell expression data of all samples in @data slot and Monocle2 object (already embedded and ordered) in @monocle_obj slot}

\item{verbose}{Boolean that determines whether progress (sequential processing of samples) should be printed. FALSE by default}

\item{cell_model}{Either "monocle2" or "seurat" depending on method used to model cell state space}
}
\value{
phemdObj with cell subtype frequencies of each sample in @data_cluster_weights slot
}
\description{
Takes as input an phemdObj with all single-cell expression data of all single-cell samples in @data slot and Monocle2 object (already embedded and ordered) in @monocle_obj slot. Returns same phemdObj with cell subtype frequencies of each sample in @data_cluster_weights slot
}
\details{
\code{embed_cells} and \code{order_cells} need to be called before calling this function.
}
\examples{

my_phemdObj <- create_dataobj(all_expn_data, all_genes, as.character(snames))
my_phemdObj_lg <- remove_tiny_samples(my_phemdObj, 10)
my_phemdObj_lg <- aggregate_samples(my_phemdObj_lg, max_cells=1000)
my_phemdObj_monocle <- embed_cells(my_phemdObj_lg, data_model = 'gaussianff', sigma=0.02, maxIter=2)
my_phemdObj_monocle <- order_cells(my_phemdObj_monocle)
my_phemdObj_final <- cluster_individual_samples(my_phemdObj_monocle)

}
