% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-core.R
\name{clusterIndividualSamples}
\alias{clusterIndividualSamples}
\title{Computes cell subtype abundances for each sample}
\usage{
clusterIndividualSamples(obj, verbose = FALSE,
  cell_model = c("monocle2", "seurat"))
}
\arguments{
\item{obj}{'Phemd' object containing single-cell expression data of all samples in @data slot and Monocle2 object (already embedded and ordered) in @monocle_obj slot}

\item{verbose}{Boolean that determines whether progress (sequential processing of samples) should be printed. FALSE by default}

\item{cell_model}{Either "monocle2" or "seurat" depending on method used to model cell state space}
}
\value{
'Phemd' object with cell subtype frequencies of each sample in @data_cluster_weights slot
}
\description{
Takes as input a Phemd object with all single-cell expression data of all single-cell samples in @data slot and Monocle2 object (already embedded and ordered) in @monocle_obj slot. Returns updated object with cell subtype frequencies of each sample in @data_cluster_weights slot
}
\details{
\code{embedCells} and \code{orderCellsMonocle} need to be called before calling this function.
}
\examples{

my_phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames))
my_phemdObj_lg <- removeTinySamples(my_phemdObj, 10)
my_phemdObj_lg <- aggregateSamples(my_phemdObj_lg, max_cells=1000)
my_phemdObj_monocle <- embedCells(my_phemdObj_lg, data_model = 'gaussianff', sigma=0.02, maxIter=2)
my_phemdObj_monocle <- orderCellsMonocle(my_phemdObj_monocle)
my_phemdObj_final <- clusterIndividualSamples(my_phemdObj_monocle)

}
