% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{compare_samples}
\alias{compare_samples}
\title{Computes EMD distance matrix representing pairwise dissimilarity between samples}
\usage{
compare_samples(obj)
}
\arguments{
\item{obj}{'phemdObj' object containing cell subtype relative frequencies for each sample in @data_cluster_weights slot and ground distance matrix (representing cell subtype dissimilarity) in @emd_dist_mat slot}
}
\value{
Distance matrix of dimension num_samples x num_samples representing pairwise dissimilarity between samples
}
\description{
Takes as input an phemdObj with cell subtype relative frequencies for each sample in @data_cluster_weights slot and ground distance matrix (representing cell subtype pairwise dissimilarity) in @emd_dist_mat slot. Returns distance matrix representing pairwise dissimilarity between samples
}
\details{
Requires 'transport' and 'pracma' packages
}
\examples{

my_phemdObj <- create_dataobj(all_expn_data, all_genes, as.character(snames))
my_phemdObj_lg <- remove_tiny_samples(my_phemdObj, 10)
my_phemdObj_lg <- aggregate_samples(my_phemdObj_lg, max_cells=1000)
my_phemdObj_monocle <- embed_cells(my_phemdObj_lg, data_model = 'gaussianff', sigma=0.02, maxIter=2)
my_phemdObj_monocle <- order_cells(my_phemdObj_monocle)
my_phemdObj_final <- cluster_individual_samples(my_phemdObj_monocle)
my_phemdObj_final <- generate_gdm(my_phemdObj_final)
my_EMD_mat <- compare_samples(my_phemdObj_final)

}
