% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{draw_colnames_45}
\alias{draw_colnames_45}
\title{Rotates heatmap marker labels 45 degrees}
\usage{
draw_colnames_45(coln, gaps, ...)
}
\arguments{
\item{coln}{column names}

\item{gaps}{spacing of labels}
}
\value{
Formatted marker labels in heatmap
}
\description{
Overwrites default draw_colnames in the pheatmap package
}
\details{
To be used with pheatmap plotting function; not to be called directly. Thanks to Josh O'Brien at http://stackoverflow.com/questions/15505607
}
\examples{
#Not to be called directly
}
