% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-plotting.R
\name{drawColnames45}
\alias{drawColnames45}
\title{Rotates heatmap marker labels 45 degrees}
\usage{
drawColnames45(coln, gaps, ...)
}
\arguments{
\item{coln}{Column names}

\item{gaps}{Spacing of labels}

\item{...}{Additional parameters to be passed to \code{gpar}}
}
\value{
Formatted marker labels in heatmap
}
\description{
Overwrites default draw_colnames in the pheatmap package
}
\details{
To be used with pheatmap plotting function; not to be called directly. Thanks to Josh O'Brien at http://stackoverflow.com/questions/15505607
}
\examples{
#Not to be called directly
}
