% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gaussianff_local}
\alias{gaussianff_local}
\title{Models expression data using generalized linear model with Gaussian error}
\usage{
gaussianff_local(dispersion = 0, parallel = FALSE, zero = NULL)
}
\arguments{
\item{dispersion}{Always set to 0 when called internally}

\item{parallel}{Always set to FALSE}

\item{zero}{Set to NULL}
}
\description{
Useful for modeling pre-normalized expression data
}
\details{
Private method (not to be called by user directly). Requires VGAM package.
}
