% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-core.R
\name{gaussianffLocal}
\alias{gaussianffLocal}
\title{Models expression data using generalized linear model with Gaussian error}
\usage{
gaussianffLocal(dispersion = 0, parallel = FALSE, zero = NULL)
}
\arguments{
\item{dispersion}{Dispersion parameter. If 0, then estimate as described in VGAM 1.0-5 documentation.}

\item{parallel}{A logical or formula. If a formula, the response of the formula should be a logical and the terms of the formula indicates whether or not those terms are parallel.}

\item{zero}{An integer-valued vector specifying which linear/additive predictors are modelled as intercepts only. The values must be from the set {1...M} where Mis the number of columns of the matrix response.}
}
\value{
Generalized linear model with Gaussian error
}
\description{
Useful for modeling pre-normalized single-cell expression data.
}
\details{
Private method (not to be called by user directly). Requires VGAM package. Obtained from VGAM v1.0-5 (https://www.rdocumentation.org/packages/VGAM/versions/1.0-5/topics/gaussianff)
}
