% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phemd-methods.R
\name{GDM}
\alias{GDM}
\title{Accessor function for EMD ground distance matrix}
\usage{
GDM(obj)
}
\arguments{
\item{obj}{A Phemd object}
}
\value{
Sqaure matrix representing pairwise distances between cell subtypes
}
\description{
Accessor function for EMD ground distance matrix
}
\examples{
phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames_data))
gdm <- GDM(phemdObj)

}
