% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_sample_sizes}
\alias{get_sample_sizes}
\title{Retrieve single-cell sample sizes}
\usage{
get_sample_sizes(data_list)
}
\arguments{
\item{data_list}{List of length num_samples (each element has dimension num_cells x num_markers)}
}
\value{
Vector of length num_samples representing number of cells in each sample
}
\description{
Takes initial list of single-cell samples and returns vector containing number of cells in each sample.
}
\details{
Private method (not exported in namespace)
}
\examples{
\dontrun{
sample_sizes <- get_sample_sizes(all_expn_data)
}
}
