% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-getResults.R
\name{getCellYield}
\alias{getCellYield}
\title{Gets cell yield of each sample as a table}
\usage{
getCellYield(myobj, cluster_assignments = NULL)
}
\arguments{
\item{myobj}{phemdObj object containing expression data for each sample in 'data' slot}

\item{cluster_assignments}{Vector of cluster assignments to be included as additional column in output table (optional)}
}
\value{
Data frame representing cell yield of each sample
}
\description{
Gets cell yield (number of viable cells) of each single-cell sample in decreasing order
}
\examples{

my_phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames_data))
my_phemdObj_lg <- removeTinySamples(my_phemdObj, 10)
my_phemdObj_lg <- aggregateSamples(my_phemdObj_lg, max_cells=1000)
my_phemdObj_monocle <- embedCells(my_phemdObj_lg, data_model = 'gaussianff', sigma=0.02, maxIter=2)
my_phemdObj_monocle <- orderCellsMonocle(my_phemdObj_monocle)
my_phemdObj_final <- clusterIndividualSamples(my_phemdObj_monocle)
my_phemdObj_final <- generateGDM(my_phemdObj_final)
my_EMD_mat <- compareSamples(my_phemdObj_final)
cluster_assignments <- groupSamples(my_EMD_mat, distfun = 'hclust', ncluster=4)
getCellYield(my_phemdObj_final, cluster_assignments)

}
