% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-getResults.R
\name{getSampleCelltypeFreqs}
\alias{getSampleCelltypeFreqs}
\title{Returns cell subtype distribution for each sample as a table}
\usage{
getSampleCelltypeFreqs(myobj, cluster_assignments = NULL)
}
\arguments{
\item{myobj}{phemdObj object containing expression data for each sample in 'data' slot}

\item{cluster_assignments}{Vector of cluster assignments to be included as additional column in output table (optional)}
}
\value{
Data frame representing relative frequencies of each cell subtype along with (optional) final inhibitor cluster assignment for each single-cell sample
}
\description{
Returns cell subtype distribution for each single-cell sample along with (optional) final inhibitor cluster assignment
}
\examples{

my_phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames_data))
my_phemdObj_lg <- removeTinySamples(my_phemdObj, 10)
my_phemdObj_lg <- aggregateSamples(my_phemdObj_lg, max_cells=1000)
my_phemdObj_monocle <- embedCells(my_phemdObj_lg, data_model = 'gaussianff', sigma=0.02, maxIter=2)
my_phemdObj_monocle <- orderCellsMonocle(my_phemdObj_monocle)
my_phemdObj_final <- clusterIndividualSamples(my_phemdObj_monocle)
my_phemdObj_final <- generateGDM(my_phemdObj_final)
my_EMD_mat <- compareSamples(my_phemdObj_final)
cluster_assignments <- groupSamples(my_EMD_mat, distfun = 'hclust', ncluster=4)
getSampleCelltypeFreqs(my_phemdObj_final, cluster_assignments)

}
