% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{group_samples}
\alias{group_samples}
\title{Performs community detection on sample-sample distance matrix}
\usage{
group_samples(distmat, distfun = "hclust", ncluster = NULL,
  method = "complete")
}
\arguments{
\item{distmat}{Distance matrix of dimension num_samples x num_samples representing pairwise dissimilarity between samples}

\item{distfun}{Method of partitioning network of samples (currently either 'hclust' or 'pam')}

\item{ncluster}{Optional parameter specifying total number of sample groups}

\item{method}{Optional parameter for hierarchical clustering (see "hclust" documentation)}
}
\value{
Vector containing group assignments for each sample (same order as row-order of distmat) based on user-specified partitioning method (e.g. hierarchical clustering)
}
\description{
Takes sample-sample distance matrix as input and returns group assignments for each sample
}
\details{
By default, uses 'kgs' (Kelley-Gardner-Sutcliffe) method for determining optimal number of groups. Alternatively, can take user-specified number of groups). Requires 'cluster' and 'maptree' packages.
}
\examples{

my_phemdObj <- create_dataobj(all_expn_data, all_genes, as.character(snames))
my_phemdObj_lg <- remove_tiny_samples(my_phemdObj, 10)
my_phemdObj_lg <- aggregate_samples(my_phemdObj_lg, max_cells=1000)
my_phemdObj_monocle <- embed_cells(my_phemdObj_lg, data_model = 'gaussianff', sigma=0.02, maxIter=2)
my_phemdObj_monocle <- order_cells(my_phemdObj_monocle)
my_phemdObj_final <- cluster_individual_samples(my_phemdObj_monocle)
my_phemdObj_final <- generate_gdm(my_phemdObj_final)
my_EMD_mat <- compare_samples(my_phemdObj_final)
cluster_assignments <- group_samples(my_EMD_mat, distfun = 'hclust', ncluster=4)

}
