% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-core.R
\name{groupSamples}
\alias{groupSamples}
\title{Performs community detection on sample-sample distance matrix}
\usage{
groupSamples(distmat, distfun = "hclust", ncluster = NULL,
  method = "complete")
}
\arguments{
\item{distmat}{Distance matrix of dimension num_samples x num_samples representing pairwise dissimilarity between samples}

\item{distfun}{Method of partitioning network of samples (currently either 'hclust' or 'pam')}

\item{ncluster}{Optional parameter specifying total number of sample groups}

\item{method}{Optional parameter for hierarchical clustering (see "hclust" documentation)}
}
\value{
Vector containing group assignments for each sample (same order as row-order of distmat) based on user-specified partitioning method (e.g. hierarchical clustering)
}
\description{
Takes sample-sample distance matrix as input and returns group assignments for each sample
}
\details{
By default, uses 'kgs' (Kelley-Gardner-Sutcliffe) method for determining optimal number of groups. Alternatively, can take user-specified number of groups). Requires 'cluster' and 'maptree' packages.
}
\examples{

my_phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames))
my_phemdObj_lg <- removeTinySamples(my_phemdObj, 10)
my_phemdObj_lg <- aggregateSamples(my_phemdObj_lg, max_cells=1000)
my_phemdObj_monocle <- embedCells(my_phemdObj_lg, data_model = 'gaussianff', sigma=0.02, maxIter=2)
my_phemdObj_monocle <- orderCellsMonocle(my_phemdObj_monocle)
my_phemdObj_final <- clusterIndividualSamples(my_phemdObj_monocle)
my_phemdObj_final <- generateGDM(my_phemdObj_final)
my_EMD_mat <- compareSamples(my_phemdObj_final)
cluster_assignments <- groupSamples(my_EMD_mat, distfun = 'hclust', ncluster=4)

}
