\name{heatmap_genes}
\alias{heatmap_genes}
\docType{data}
\title{Genes to be used when plotting heatmap for melanoma single-cell RNA-seq expression data}
\description{This object contains genes to be used when plotting heatmap for melanoma single-cell RNA-seq expression data.}
\usage{
  data(heatmap_genes)
}
\format{
  Vector of length 42 representing selected genes for plotting heatmap.
}
\source{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE72056}
\references{Tirosh, I. et al. Dissecting the multicellular ecosystem of metastatic melanoma by single-cell RNA-seq. Science 352, 189–196 (2016)}
\keyword{datasets}
