% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{identify_centroids}
\alias{identify_centroids}
\title{Identify cluster centroids (cell names)}
\usage{
identify_centroids(ref_clusters, data_ordered)
}
\arguments{
\item{ref_clusters}{list containing each cluster of interest (each list element is a matrix of dimension num_cells x num_markers)}
}
\value{
List of names; element \var{i} represents the name of the cell in cluster \var{i} that is closest to the centroid (arithmetic mean) of cluster \var{i}
}
\description{
Takes initial list and returns list of cell names representing centroid of cluster
}
\details{
Private method (not exported in namespace)
}
\examples{
\dontrun{
centroid_names <- identify_centroids(ref_clusters, data_ordered)
}
}
