% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-core.R
\name{identifyCentroids}
\alias{identifyCentroids}
\title{Identify cluster centroids (cell names)}
\usage{
identifyCentroids(ref_clusters)
}
\arguments{
\item{ref_clusters}{list containing each cluster of interest (each list element is a matrix of dimension num_cells x num_markers)}
}
\value{
List of names; element \var{i} represents the name of the cell in cluster \var{i} that is closest to the centroid (arithmetic mean) of cluster \var{i}
}
\description{
Takes initial list and returns list of cell names representing centroid of cluster
}
\details{
Private method (not exported in namespace)
}
\examples{
\dontrun{
centroid_names <- identifyCentroids(ref_clusters)
}
}
