% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phemd-methods.R
\name{monocleInfo}
\alias{monocleInfo}
\title{Accessor function for stored Monocle object}
\usage{
monocleInfo(obj)
}
\arguments{
\item{obj}{A Phemd object.}
}
\value{
An object of class 'CellDataSet' (from Monocle)
}
\description{
Accessor function for stored Monocle object
}
\examples{
phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames_data))
monocle_obj <- monocleInfo(phemdObj)

}
