% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-core.R
\name{orderCellsMonocle}
\alias{orderCellsMonocle}
\title{Compute Monocle2 cell state and pseudotime assignments}
\usage{
orderCellsMonocle(obj, ...)
}
\arguments{
\item{obj}{'Phemd' object containing initial Monocle 2 object}

\item{...}{Additional parameters to be passed into \code{orderCells} function}
}
\value{
Same as input 'Phemd' object with updated Monocle2 object in @monocle_obj slot containing cell state and pseudotime assignments
}
\description{
Takes as input a Phemd object with Monocle2 object and returns updated object with Monocle2 object containing cell state and pseudotime assignments
}
\details{
Wrapper function for \code{orderCells} in Monocle 2 package. \code{embedCells} needs to be called before calling this function.
}
\examples{

my_phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames_data))
my_phemdObj_lg <- removeTinySamples(my_phemdObj, 10)
my_phemdObj_lg <- aggregateSamples(my_phemdObj_lg, max_cells=1000)
my_phemdObj_monocle <- embedCells(my_phemdObj_lg, data_model='gaussianff', sigma=0.02, maxIter=2)
my_phemdObj_monocle <- orderCellsMonocle(my_phemdObj_monocle)
}
