% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phemd-methods.R
\name{phateInfo}
\alias{phateInfo}
\title{Accessor function for stored phate object}
\usage{
phateInfo(obj)
}
\arguments{
\item{obj}{A Phemd object.}
}
\value{
An object of class 'phate' (from phateR)
}
\description{
Accessor function for stored phate object
}
\examples{
phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames_data))
phateobj <- phateInfo(phemdObj)

}
