% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phemd-methods.R
\docType{methods}
\name{Phemd-methods}
\alias{Phemd-methods}
\alias{selectMarkers<-}
\alias{selectMarkers<-,Phemd-method}
\alias{Phemd,ANY,ANY-method}
\alias{rawExpn<-}
\alias{Phemd,character,ANY-method}
\alias{rawExpn<-,Phemd-method}
\alias{pooledCells<-}
\alias{pooledCells<-,Phemd-method}
\alias{subsampledIdx<-}
\alias{subsampledIdx<-,Phemd-method}
\alias{subsampledBool<-}
\alias{subsampledBool<-,Phemd-method}
\alias{monocleInfo<-}
\alias{monocleInfo<-,Phemd-method}
\alias{seuratInfo<-}
\alias{seuratInfo<-,Phemd-method}
\alias{celltypeFreqs<-}
\alias{celltypeFreqs<-,Phemd-method}
\alias{batchIDs<-}
\alias{batchIDs<-,Phemd-method}
\alias{GDM<-}
\alias{GDM<-,Phemd-method}
\title{Setter function for protein / gene markers}
\usage{
selectMarkers(obj) <- value

\S4method{selectMarkers}{Phemd}(obj) <- value

rawExpn(obj) <- value

\S4method{rawExpn}{Phemd}(obj) <- value

pooledCells(obj) <- value

\S4method{pooledCells}{Phemd}(obj) <- value

subsampledIdx(obj) <- value

\S4method{subsampledIdx}{Phemd}(obj) <- value

subsampledBool(obj) <- value

\S4method{subsampledBool}{Phemd}(obj) <- value

monocleInfo(obj) <- value

\S4method{monocleInfo}{Phemd}(obj) <- value

seuratInfo(obj) <- value

\S4method{seuratInfo}{Phemd}(obj) <- value

celltypeFreqs(obj) <- value

\S4method{celltypeFreqs}{Phemd}(obj) <- value

batchIDs(obj) <- value

\S4method{batchIDs}{Phemd}(obj) <- value

GDM(obj) <- value

\S4method{GDM}{Phemd}(obj) <- value
}
\arguments{
\item{obj}{A Phemd object}

\item{value}{Assignment object}
}
\value{
Updated Phemd object

Updated Phemd object

Updated Phemd object

Updated Phemd object

Updated Phemd object

Updated Phemd object containing Seurat object

Updated Phemd object

Updated Phemd object

Updated Phemd object
}
\description{
Setter function for protein / gene markers

Setter function for stored expression data

Setter function for single-cell expression data aggregated from multiple samples

Setter function for indices of cells subsampled from each sample during aggregation

Setter function for boolean denoting whether cells were subsampled from each sample during aggregation

Setter function for Monocle2 CellDataSet object for experiment

Setter function for Seurat object for experiment

Setter function for cell subtype frequencies of each single-cell sample

Setter function for batch IDs of each single-cell sample

Setter function for EMD ground distance matrix
}
\examples{
\dontrun{
selectMarkers(phemdObj) <- new_markers
}
\dontrun{
rawExpn(phemdObj) <- new_expn_data
}
\dontrun{
pooledCells(phemdObj) <- new_aggregated_data
}
\dontrun{
subsampledIdx(phemdObj) <- subsampledIdxList
}
\dontrun{
subsampledBool(phemdObj) <- TRUE
}
\dontrun{
monocleInfo(phemdObj) <- myCellDataSet
}
\dontrun{
seuratInfo(phemdObj) <- mySeuratObj
}
\dontrun{
celltypeFreqs(phemdObj) <- myCellTypeFreqs
}
\dontrun{
batchIDs(phemdObj) <- c('A', 'A', 'A', 'B', 'B')
}
\dontrun{
GDM(phemdObj) <- myGDM
}
}
