% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phemd-methods.R
\docType{class}
\name{Phemd}
\alias{Phemd}
\alias{Phemd-class}
\title{Phemd class}
\description{
The main PhEMD class to store single-cell expression data.
}
\section{Fields}{

\describe{
\item{\code{data}}{List of matrices, each of which represents a single-cell sample (num_cells x num_genes)}

\item{\code{markers}}{Column names (e.g. genes) for each element (i.e. data matrix) in "data"}

\item{\code{snames}}{Sample ID for each element in "data"}

\item{\code{data_aggregate}}{Numeric matrix representing expression data for cells from all experimental conditions (rows = markers, cols = cells)}

\item{\code{data_subsample_idx}}{List of vectors each representing the indices of elements in "data" that were subsampled and combined to form "data_aggregate"}

\item{\code{subsampled_bool}}{Boolean represent whether or not subsampling was performed in the data aggregation process}

\item{\code{monocle_obj}}{Data object of type "CellDataSet" that is the core Monocle data structure}

\item{\code{data_cluster_weights}}{Matrix representing cell subtype relative frequencies for each sample (num_samples x num_genes)}

\item{\code{emd_dist_mat}}{Matrix representing pairwise distances between each pair of cell subtypes}

\item{\code{seurat_obj}}{Object of class "Seurat" that is the core Seurat data structure}

\item{\code{phate_obj}}{Object of class "phate" that is the core PHATE data structure}

\item{\code{experiment_ids}}{Vector of length num_samples representing the experiment (batch) in which the sample was profiled}
}}

