% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_sample_histograms}
\alias{plot_sample_histograms}
\title{Plot cell subtype frequency histograms for each sample}
\usage{
plot_sample_histograms(myobj, cluster_assignments, dest, cmap = NULL)
}
\arguments{
\item{myobj}{phemdObj object containing cell subtype relative frequency in @data_cluster_weights slot}

\item{cluster_assignments}{Vector containing group assignments for each sample in myobj}

\item{dest}{Path to existing directory where output should be saved}

\item{cmap}{Vector containing colors by which histogram bars should be colored (optional)}
}
\value{
None
}
\description{
Plots relative frequency ("weights") of cell subtypes ("bins" or "signatures") in each single-cell sample
}
\details{
\code{group_samples} must be called before calling this function. Saves plots in directory called "individual_inhibs"
}
\examples{
my_phemdObj_monocle <- embed_cells(my_phemdObj_lg, data_model = 'gaussianff', sigma=0.02)
my_phemdObj_monocle <- order_cells(my_phemdObj_monocle)
cmap <- plot_embeddings(my_phemdObj_monocle, savedest)
my_phemdObj_final <- cluster_individual_samples(my_phemdObj_monocle)
my_phemdObj_final <- generate_gdm(my_phemdObj_final)
my_EMD_mat <- compare_samples(my_phemdObj_final)
cluster_assignments <- group_samples(my_EMD_mat, distfun = 'hclust')
print_cluster_assignments(cluster_assignments, myobj, savedest)
plot_sample_histograms(my_phemdObj_final, cluster_assignments, savedest, cmap)

}
