% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-plotting.R
\name{plotEmbeddings}
\alias{plotEmbeddings}
\title{Plots Monocle2 cell embedding plots}
\usage{
plotEmbeddings(obj, cell_model = c("monocle2", "seurat"), cmap = NULL,
  w = 4, h = 5, pt_sz = 1, ndims = NULL)
}
\arguments{
\item{obj}{'Phemd' object containing Monocle 2 object}

\item{cell_model}{Method by which cell state was modeled (either "monocle2" or "seurat")}

\item{cmap}{User-specified colormap to use to color cell state embedding (optional)}

\item{w}{Width of plot in inches}

\item{h}{Height of plot in inches}

\item{pt_sz}{Scalar factor for point size}

\item{ndims}{Number of dimensions to use for dimensionality reduction in case it hasn't been performed yet (only relevant when using Seurat data as input)}
}
\value{
Colormap (vector of colors) used to color Monocle2 cell state embedding
}
\description{
Takes as input a Phemd object containing either a Monocle2 object or Seurat object (already embedded and ordered) and plots cell embedding plots side by side. Optionally saves to specified folder.
}
\details{
\code{embedCells} and \code{orderCellsMonocle} need to be called before calling this function. Required additional packages: 'RColorBrewer', 'cowplot'
}
\examples{
my_phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames_data))
my_phemdObj_lg <- removeTinySamples(my_phemdObj, 10)
my_phemdObj_lg <- aggregateSamples(my_phemdObj_lg, max_cells=1000)
my_phemdObj_monocle <- embedCells(my_phemdObj_lg, data_model='gaussianff', sigma=0.02, maxIter=2)
my_phemdObj_monocle <- orderCellsMonocle(my_phemdObj_monocle)
cmap <- plotEmbeddings(my_phemdObj_monocle)
}
