% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-plotting.R
\name{plotSummaryHistograms}
\alias{plotSummaryHistograms}
\title{Plots cell subtype frequency histograms summarizing each group of samples}
\usage{
plotSummaryHistograms(
  myobj,
  cluster_assignments,
  cell_model = c("monocle2", "seurat", "phate"),
  cmap = NULL,
  ncol.plot = 4,
  ax.lab.sz = 2.5,
  title.sz = 3
)
}
\arguments{
\item{myobj}{Phemd object containing cell subtype relative frequency in @data_cluster_weights slot}

\item{cluster_assignments}{Vector containing group assignments for each sample in myobj}

\item{cell_model}{Method by which cell state was modeled (either "monocle2", "seurat", or "phate")}

\item{cmap}{Vector containing colors by which histogram bars should be colored (optional)}

\item{ncol.plot}{Number of columns to use to plot multi-panel histogram plot}

\item{ax.lab.sz}{Scaling factor for axis labels (default 2.5)}

\item{title.sz}{Scaling factor for plot title (default 3)}
}
\value{
None
}
\description{
Visualizes plots of relative frequency ("weights") of cell subtypes ("bins" or "signatures") summarizing each group of single-cell samples. Each summary histogram is computed by taking the bin-wise mean of all samples in the group
}
\details{
\code{groupSamples} must be called before calling this function. Saves plots in directory called "summary_inhibs"
}
