% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phemd-methods.R
\name{pooledCells}
\alias{pooledCells}
\title{Accessor function for aggregated cells used for cell subtype definition}
\usage{
pooledCells(obj)
}
\arguments{
\item{obj}{Phemd object}
}
\value{
Numeric matrix representing expression data for cells from all experimental conditions (rows = markers, cols = cells)
}
\description{
Accessor function for aggregated cells used for cell subtype definition
}
\examples{
phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames))
cells_aggregated <- pooledCells(phemdObj)

}
