% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print_cluster_assignments}
\alias{print_cluster_assignments}
\title{Writes samples to file based on community detection group assignments}
\usage{
print_cluster_assignments(cluster_assignments, obj, dest)
}
\arguments{
\item{cluster_assignments}{Vector containing group assignments for each sample}

\item{obj}{phemdObj object containing sample names in @snames slot}

\item{dest}{Path to existing directory where output should be saved}
}
\value{
None
}
\description{
Takes vector of cluster assignments and phemdObj containing sample names and writes sample groups to file
}
\details{
Order of samples in obj@snames is assumed to be the same as the order of group assignments in cluster_assignments
}
\examples{

my_phemdObj <- create_dataobj(all_expn_data, all_genes, as.character(snames))
my_phemdObj_lg <- remove_tiny_samples(my_phemdObj, 10)
my_phemdObj_lg <- aggregate_samples(my_phemdObj_lg, max_cells=1000)
my_phemdObj_monocle <- embed_cells(my_phemdObj_lg, data_model = 'gaussianff', sigma=0.02, maxIter=2)
my_phemdObj_monocle <- order_cells(my_phemdObj_monocle)
my_phemdObj_final <- cluster_individual_samples(my_phemdObj_monocle)
my_phemdObj_final <- generate_gdm(my_phemdObj_final)
my_EMD_mat <- compare_samples(my_phemdObj_final)
cluster_assignments <- group_samples(my_EMD_mat, distfun = 'hclust', ncluster=4)
print_cluster_assignments(cluster_assignments, my_phemdObj_final, '.')

}
