% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phemd-methods.R
\name{rawExpn}
\alias{rawExpn}
\title{Accessor function for stored multi-sample raw expression data}
\usage{
rawExpn(obj)
}
\arguments{
\item{obj}{A Phemd object.}
}
\value{
List of matrices, each of which represents a single-cell sample
}
\description{
Accessor function for stored multi-sample raw expression data
}
\examples{
phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames))
raw_expn_data <- rawExpn(phemdObj)

}
