% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{remove_tiny_samples}
\alias{remove_tiny_samples}
\title{Remove samples with too few cells}
\usage{
remove_tiny_samples(obj, min_sz = 20)
}
\arguments{
\item{obj}{'phemdObj' object containing raw expression data and associated metadata}

\item{min_sz}{Minimum number of cells in each sample to be retained}
}
\value{
'phemdObj' object containing raw multi-sample expression data and associated metadata (same as input minus removed samples)
}
\description{
Removes samples from phemdObj that have fewer cells than \code{min_sz}
}
\details{
Note: If used, this function must be called before (and not after) the \code{aggregate_samples} function is called
}
\examples{
\dontrun{
my_phemdObj <- create_dataobj(all_expn_data, genes_measured, my_sample_names)
my_phemdObj_lg <- remove_tiny_samples(my_phemdObj, 10) #removes samples with fewer than 10 cells
}
}
