% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-core.R
\name{removeTinySamples}
\alias{removeTinySamples}
\title{Remove samples with too few cells}
\usage{
removeTinySamples(obj, min_sz = 20)
}
\arguments{
\item{obj}{'phemdObj' object containing raw expression data and associated metadata}

\item{min_sz}{Minimum number of cells in each sample to be retained}
}
\value{
'phemdObj' object containing raw multi-sample expression data and associated metadata (same as input minus removed samples)
}
\description{
Removes samples from phemdObj that have fewer cells than \code{min_sz}
}
\details{
Note: If used, this function must be called before (and not after) the \code{aggregateSamples} function is called
}
\examples{
my_phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames))
my_phemdObj_lg <- removeTinySamples(my_phemdObj, 10) #removes samples with fewer than 10 cells

}
