% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-saveResults.R
\name{saveSampleHistograms}
\alias{saveSampleHistograms}
\title{Saves cell subtype frequency histograms for each sample}
\usage{
saveSampleHistograms(myobj, cluster_assignments, dest,
  cell_model = c("monocle2", "seurat"), cmap = NULL,
  overwrite = FALSE)
}
\arguments{
\item{myobj}{phemdObj object containing cell subtype relative frequency in @data_cluster_weights slot}

\item{cluster_assignments}{Vector containing group assignments for each sample in myobj}

\item{dest}{Path to existing directory where output should be saved}

\item{cell_model}{Method by which cell state was modeled (either "monocle2" or "seurat")}

\item{cmap}{Vector containing colors by which histogram bars should be colored (optional)}

\item{overwrite}{Boolean representing whether or not to overwrite contents of "dest" with output of saveSampleHistograms}
}
\value{
None
}
\description{
Saves relative frequency ("weights") of cell subtypes ("bins" or "signatures") in each single-cell sample
}
\details{
\code{groupSamples} must be called before calling this function. Saves plots in directory called "individual_inhibs"
}
\examples{

my_phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames))
my_phemdObj_lg <- removeTinySamples(my_phemdObj, 10)
my_phemdObj_lg <- aggregateSamples(my_phemdObj_lg, max_cells=1000)
my_phemdObj_monocle <- embedCells(my_phemdObj_lg, data_model = 'gaussianff', sigma=0.02, maxIter=2)
my_phemdObj_monocle <- orderCellsMonocle(my_phemdObj_monocle)
my_phemdObj_final <- clusterIndividualSamples(my_phemdObj_monocle)
my_phemdObj_final <- generateGDM(my_phemdObj_final)
my_EMD_mat <- compareSamples(my_phemdObj_final)
cluster_assignments <- groupSamples(my_EMD_mat, distfun = 'hclust', ncluster=4)
printClusterAssignments(cluster_assignments, my_phemdObj_final, '.', overwrite=TRUE)
dm <- plotGroupedSamplesDmap(my_EMD_mat, cluster_assignments, dest=NULL, pt_sz=2)
saveSampleHistograms(my_phemdObj_final, cluster_assignments, '.', overwrite=TRUE)

}
