% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{select_features}
\alias{select_features}
\title{Perform feature selection on aggregated data}
\usage{
select_features(obj, selected_genes)
}
\arguments{
\item{obj}{'phemdObj' object containing aggregated data}

\item{selected_genes}{Vector containing names of genes to use for downstream analyses}
}
\value{
Same as input 'phemdObj' object after performing feature-selection based dimensionality reduction on aggregated expression data
}
\description{
Takes as input an phemdObj with aggregated data and returns same phemdObj after performing feature selection on aggregated data
}
\details{
\code{aggregate_samples} needs to be called before running this function
}
\examples{

my_phemdObj <- create_dataobj(all_expn_data, all_genes, as.character(snames))
my_phemdObj_lg <- remove_tiny_samples(my_phemdObj, 10)
my_phemdObj_lg <- aggregate_samples(my_phemdObj_lg, max_cells=1000)
my_phemdObj_lg <- select_features(my_phemdObj_lg, selected_genes=c('TP53', 'EGFR', 'KRAS'))

}
