\name{selected_genes}
\alias{selected_genes}
\docType{data}
\title{Genes to be used when performing clustering and trajectory analyses on melanoma single-cell RNA-seq expression data}
\description{This object contains genes to be used when performing clustering and trajectory analyses on melanoma single-cell RNA-seq expression data.}
\usage{
  data(selected_genes)
}
\format{
  Vector of length 44 representing selected genes for performing computational analyses such as generating cell embeddings and clustering cell subtypes.
}
\source{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE72056}
\references{Tirosh, I. et al. Dissecting the multicellular ecosystem of metastatic melanoma by single-cell RNA-seq. Science 352, 189–196 (2016)}
\keyword{datasets}
