% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phemd-methods.R
\name{selectMarkers}
\alias{selectMarkers}
\title{Accessor function for gene/protein markers measured in experiment}
\usage{
selectMarkers(obj)
}
\arguments{
\item{obj}{Phemd object}
}
\value{
Vector representing gene/protein markers corresponding to expression matrices
}
\description{
Accessor function for gene/protein markers measured in experiment
}
\examples{
phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames))
genes <- selectMarkers(phemdObj)

}
