% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phemd-methods.R
\name{seuratInfo}
\alias{seuratInfo}
\title{Accessor function for stored Seurat object within Phemd object}
\usage{
seuratInfo(obj)
}
\arguments{
\item{obj}{A Phemd object.}
}
\value{
An object of class 'Seurat'
}
\description{
Accessor function for stored Seurat object within Phemd object
}
\examples{
phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames_data))
seurat_obj <- seuratInfo(phemdObj)

}
