% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phemd-methods.R
\name{sNames}
\alias{sNames}
\title{Accessor function for identifiers of all single-cell samples in experiment}
\usage{
sNames(obj)
}
\arguments{
\item{obj}{Phemd object}
}
\value{
Vector representing sample names corresponding to expression matrices
}
\description{
Accessor function for identifiers of all single-cell samples in experiment
}
\examples{
phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames_data))
sampleIDs <- sNames(phemdObj)

}
