\name{snames_data}
\alias{snames_data}
\docType{data}
\title{Sample names for melanoma single-cell RNA-seq expression data}
\description{This object contains sample names corresponding to samples contained in melanoma expression data.}
\usage{
  data("snames_data")
  }
\format{
  Vector of length 19 representing sample names corresponding to order of samples in all_expn_data in melanomaData dataset.
}
\source{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE72056}
\references{Tirosh, I. et al. Dissecting the multicellular ecosystem of metastatic melanoma by single-cell RNA-seq. Science 352, 189–196 (2016)}
\keyword{datasets}
