% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phemd-methods.R
\name{subsampledBool}
\alias{subsampledBool}
\title{Accessor function for whether or not cells were subsampled when aggregated for cell subtype analysis}
\usage{
subsampledBool(obj)
}
\arguments{
\item{obj}{Phemd object}
}
\value{
Boolean represent whether or not subsampling was performed in the data aggregation process
}
\description{
Accessor function for whether or not cells were subsampled when aggregated for cell subtype analysis
}
\examples{
phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames))
subsampled <- subsampledBool(phemdObj)

}
