% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-core.R
\name{bindSeuratObj}
\alias{bindSeuratObj}
\title{Attach 'Seurat' object to 'Phemd' object}
\usage{
bindSeuratObj(phemd_obj, seurat_obj, batch.colname = "plt")
}
\arguments{
\item{phemd_obj}{Phemd object initialized using createDataObj}

\item{seurat_obj}{S4 'seurat' object containing batch-normalized reference cell data}

\item{batch.colname}{Name of column in Seurat object that denotes batch ID}
}
\value{
'Phemd' object containing with attached Seurat object
}
\description{
Allows user to attach batch-normalized reference cell data from Seurat into 'Phemd' object containing raw expression data and metadata
}
\examples{

my_phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames_data))
my_seuratObj <- Seurat::CreateSeuratObject(counts = t(all_expn_data[[1]]), project = "A")
my_seuratObj <- Seurat::FindVariableFeatures(object = my_seuratObj)
my_seuratObj <- Seurat::ScaleData(object = my_seuratObj, do.scale=FALSE, do.center=FALSE)
my_seuratObj <- Seurat::RunPCA(object = my_seuratObj, pc.genes = colnames(all_expn_data[[1]]), do.print = FALSE)
my_seuratObj <- Seurat::FindNeighbors(my_seuratObj, reduction = "pca", dims.use = 1:10)
my_seuratObj <- Seurat::FindClusters(my_seuratObj, resolution = 0.6, print.output = 0, save.SNN = TRUE)
my_phemdObj <- bindSeuratObj(my_phemdObj, my_seuratObj)

}
