% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{generate_gdm}
\alias{generate_gdm}
\title{Computes ground distance matrix based on cell embedding}
\usage{
generate_gdm(obj, cell_model = "monocle2")
}
\arguments{
\item{obj}{'phemdObj' object containing Monocle2 object (already embedded and ordered) in @monocle_obj slot}

\item{cell_model}{Method by which cell state was modeled (either "monocle2" or "seurat")}
}
\value{
phemdObj with ground distance matrix (to be used in EMD computation) in @data_cluster_weights slot
}
\description{
Takes as input an phemdObj with Monocle2 object (already embedded and ordered) in @monocle_obj slot. Returns same phemdObj with ground distance matrix representing pairwise distance between 2 cell subtypes based on cell state embedding.
}
\details{
\code{embed_cells} and \code{order_cells} need to be called before calling this function. Requires 'igraph' package
}
\examples{

my_phemdObj <- create_dataobj(all_expn_data, all_genes, as.character(snames))
my_phemdObj_lg <- remove_tiny_samples(my_phemdObj, 10)
my_phemdObj_lg <- aggregate_samples(my_phemdObj_lg, max_cells=1000)
my_phemdObj_monocle <- embed_cells(my_phemdObj_lg, data_model = 'gaussianff', sigma=0.02, maxIter=2)
my_phemdObj_monocle <- order_cells(my_phemdObj_monocle)
my_phemdObj_final <- cluster_individual_samples(my_phemdObj_monocle)
my_phemdObj_final <- generate_gdm(my_phemdObj_final)

}
