% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-core.R
\name{generateGDM}
\alias{generateGDM}
\title{Computes ground distance matrix based on cell embedding}
\usage{
generateGDM(obj, cell_model = c("monocle2", "seurat"))
}
\arguments{
\item{obj}{'Phemd' object containing Monocle2 object (already embedded and ordered) in @monocle_obj slot}

\item{cell_model}{Method by which cell state was modeled (either "monocle2" or "seurat")}
}
\value{
Phemd object with ground distance matrix (to be used in EMD computation) in @data_cluster_weights slot
}
\description{
Takes as input a Phemd object with Monocle2 object (already embedded and ordered) in @monocle_obj slot. Returns updated object with ground distance matrix representing pairwise distance between 2 cell subtypes based on cell state embedding.
}
\details{
\code{embedCells} and \code{orderCellsMonocle} need to be called before calling this function. Requires 'igraph' package
}
\examples{

my_phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames_data))
my_phemdObj_lg <- removeTinySamples(my_phemdObj, 10)
my_phemdObj_lg <- aggregateSamples(my_phemdObj_lg, max_cells=1000)
my_phemdObj_monocle <- embedCells(my_phemdObj_lg, data_model = 'gaussianff', sigma=0.02, maxIter=2)
my_phemdObj_monocle <- orderCellsMonocle(my_phemdObj_monocle)
my_phemdObj_final <- clusterIndividualSamples(my_phemdObj_monocle)
my_phemdObj_final <- generateGDM(my_phemdObj_final)

}
