% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_cell_yield}
\alias{plot_cell_yield}
\title{Plot cell yield of each sample as bar plot}
\usage{
plot_cell_yield(myobj, dest, labels = NULL, cmap = NULL,
  font_sz = 0.6, w = 8, h = 9.5)
}
\arguments{
\item{myobj}{phemdObj object containing expression data for each sample in 'data' slot}

\item{dest}{Path to existing directory where output should be saved}

\item{labels}{Vector containing group labels for samples (optional). If not provided, bars will be of uniform color (blue)}

\item{cmap}{Vector containing colors by which histogram bars should be colored (optional)}

\item{font_sz}{Scaling factor for font size of sample names in barplot}
}
\value{
None
}
\description{
Plots cell yield (number of viable cells) of each single-cell sample in decreasing order as horizontal bar plot
}
\examples{
my_phemdObj_monocle <- embed_cells(my_phemdObj_lg, data_model = 'gaussianff', sigma=0.02)
my_phemdObj_monocle <- order_cells(my_phemdObj_monocle)
cmap <- plot_embeddings(my_phemdObj_monocle, savedest)
my_phemdObj_final <- cluster_individual_samples(my_phemdObj_monocle)
my_phemdObj_final <- generate_gdm(my_phemdObj_final)
my_EMD_mat <- compare_samples(my_phemdObj_final)
cluster_assignments <- group_samples(my_EMD_mat, distfun = 'hclust')
print_cluster_assignments(cluster_assignments, myobj, savedest)
plot_sample_histograms(my_phemdObj_final, cluster_assignments, savedest, cmap)
plot_cell_yield(myobj, savedest, cluster_assignments, font_sz = 0.8)

}
