% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_embeddings}
\alias{plot_embeddings}
\title{Save Monocle2 cell embedding plots to folder}
\usage{
plot_embeddings(obj, path, cell_model = "monocle2", cmap = NULL,
  w = 4, h = 5, pt_sz = 1, ndims = NULL)
}
\arguments{
\item{obj}{'phemdObj' object containing Monocle 2 object}

\item{path}{Path to destination folder (must already exist)}

\item{cell_model}{Method by which cell state was modeled (either "monocle2" or "seurat")}

\item{cmap}{User-specified colormap to use to color cell state embedding (optional)}
}
\value{
Colormap (vector of colors) used to color Monocle2 cell state embedding
}
\description{
Takes as input an phemdObj with Monocle2 object (already embedded and ordered) and saves cell embedding plots to specified folder
}
\details{
\code{embed_cells} and \code{order_cells} need to be called before calling this function. Required additional package: 'RColorBrewer'
}
\examples{
\dontrun{
my_phemdObj_lg <- remove_tiny_samples(my_phemdObj, 10) #removes samples with fewer than 10 cells
my_phemdObj_lg <- aggregate_samples(my_phemdObj_lg, max_cells=10000)
my_phemdObj_lg <- select_features(my_phemdObj_lg, selected_genes=c('TP53', 'EGFR', 'KRAS'))
my_phemdObj_monocle <- embed_cells(my_phemdObj_lg, data_model = 'gaussianff', sigma=0.02)
my_phemdObj_monocle <- order_cells(my_phemdObj_monocle)
cmap <- plot_embeddings(my_phemdObj_monocle, savedest)
}
}
