% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_grouped_samples_dmap}
\alias{plot_grouped_samples_dmap}
\title{Plot diffusion map embedding of samples based on distance matrix}
\usage{
plot_grouped_samples_dmap(my_distmat, cluster_assignments, dest,
  pt_sz = 1, n_dim = 3, pt_label = NULL, cmap = NULL, w = 8,
  h = 5, scale.y = 1, angle = 40, ...)
}
\arguments{
\item{my_distmat}{phemdObj object containing sample names in @snames slot}

\item{cluster_assignments}{Vector containing group assignments for each sample}

\item{dest}{Path to existing directory where output should be saved}

\item{pt_sz}{Size of points representing samples in plot (scaling factor)}

\item{n_dim}{Number of dimensions for embedding (either 2 or 3)}

\item{pt_label}{Vector of sample names corresponding to each point (same order as samples in \code{my_distmat} and \code{cluster_assignments})}

\item{cmap}{Vector containing colors by which points should be colored (corresponding to cluster_assignments)}

\item{w}{Width of plot in inches}

\item{h}{Height of plot in inches}

\item{scale.y}{Scaling factor for diffusion map y-axis}

\item{angle}{Rotation factor for diffusion map plot}

\item{...}{Additional parameters to be passed to \code{DiffusionMap} function}
}
\value{
DiffusionMap object containing biological sample embedding and associated metadata
}
\description{
Visualizes diffusion map for network of samples based on square distance matrix (sample-sample pairwise dissimilarity)
}
\details{
Requires 'destiny' package
}
\examples{

my_phemdObj <- create_dataobj(all_expn_data, all_genes, as.character(snames))
my_phemdObj_lg <- remove_tiny_samples(my_phemdObj, 10)
my_phemdObj_lg <- aggregate_samples(my_phemdObj_lg, max_cells=1000)
my_phemdObj_monocle <- embed_cells(my_phemdObj_lg, data_model = 'gaussianff', sigma=0.02, maxIter=2)
my_phemdObj_monocle <- order_cells(my_phemdObj_monocle)
my_phemdObj_final <- cluster_individual_samples(my_phemdObj_monocle)
my_phemdObj_final <- generate_gdm(my_phemdObj_final)
my_EMD_mat <- compare_samples(my_phemdObj_final)
cluster_assignments <- group_samples(my_EMD_mat, distfun = 'hclust', ncluster=4)
print_cluster_assignments(cluster_assignments, my_phemdObj_final, '.')
dm <- plot_grouped_samples_dmap(my_EMD_mat, cluster_assignments, '.', pt_sz=2, pt_label = my_phemdObj_final@snames)

}
