% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-plotting.R
\name{plotSummaryHistograms}
\alias{plotSummaryHistograms}
\title{Saves cell subtype frequency histograms summarizing each group of samples}
\usage{
plotSummaryHistograms(myobj, cluster_assignments, dest = NULL,
  cell_model = c("monocle2", "seurat"), cmap = NULL,
  overwrite = FALSE, ncol.plot = 4, ax.lab.sz = 2.5, title.sz = 3)
}
\arguments{
\item{myobj}{phemdObj object containing cell subtype relative frequency in @data_cluster_weights slot}

\item{cluster_assignments}{Vector containing group assignments for each sample in myobj}

\item{dest}{Path to existing directory where output should be saved}

\item{cell_model}{Method by which cell state was modeled (either "monocle2" or "seurat")}

\item{cmap}{Vector containing colors by which histogram bars should be colored (optional)}

\item{overwrite}{Boolean representing whether or not to overwrite contents of "dest" with output of saveSummaryHistograms}

\item{ncol.plot}{Number of columns to use to plot multi-panel histogram plot}

\item{ax.lab.sz}{Scaling factor for axis labels (default 2.5)}

\item{title.sz}{Scaling factor for plot title (default 3)}
}
\value{
None
}
\description{
Saves plots of relative frequency ("weights") of cell subtypes ("bins" or "signatures") summarizing each group of single-cell samples. Each summary histogram is computed by taking the bin-wise mean of all samples in the group
}
\details{
\code{groupSamples} must be called before calling this function. Saves plots in directory called "summary_inhibs"
}
\examples{

my_phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames))
my_phemdObj_lg <- removeTinySamples(my_phemdObj, 10)
my_phemdObj_lg <- aggregateSamples(my_phemdObj_lg, max_cells=1000)
my_phemdObj_monocle <- embedCells(my_phemdObj_lg, data_model = 'gaussianff', sigma=0.02, maxIter=2)
my_phemdObj_monocle <- orderCellsMonocle(my_phemdObj_monocle)
my_phemdObj_final <- clusterIndividualSamples(my_phemdObj_monocle)
my_phemdObj_final <- generateGDM(my_phemdObj_final)
my_EMD_mat <- compareSamples(my_phemdObj_final)
cluster_assignments <- groupSamples(my_EMD_mat, distfun = 'hclust', ncluster=4)
printClusterAssignments(cluster_assignments, my_phemdObj_final, '.', overwrite=TRUE)
dm <- plotGroupedSamplesDmap(my_EMD_mat, cluster_assignments, '.', pt_sz=2, pt_label = sampleNames(my_phemdObj_final))
plotSummaryHistograms(my_phemdObj_final, cluster_assignments)

}
