% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{retrieve_refclusters}
\alias{retrieve_refclusters}
\title{Retrieve reference cell clusters}
\usage{
retrieve_refclusters(obj)
}
\arguments{
\item{obj}{phemdObj struct containing Monocle2 object and underlying expression data}
}
\value{
List of data matrices; each list element is of size num_cells_in_cluster x num_markers and represents a distinct cell cluster
}
\description{
Takes initial phemdObj struct and returns cell clusters as assigned by clustering algorithm (i.e. Monocle 2)
}
\details{
Private method (not exported in namespace)
}
\examples{
cluster_expression_data <- retrieve_reference_clusters(my_phemdObj)
}
