% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-core.R
\name{selectFeatures}
\alias{selectFeatures}
\title{Perform feature selection on aggregated data}
\usage{
selectFeatures(obj, selected_genes)
}
\arguments{
\item{obj}{'Phemd' object containing aggregated data}

\item{selected_genes}{Vector containing names of genes to use for downstream analyses}
}
\value{
Same as input 'Phemd' object after performing feature-selection based dimensionality reduction on aggregated expression data
}
\description{
Takes as input a Phemd object with aggregated data and returns updated object after performing feature selection on aggregated data
}
\details{
\code{aggregateSamples} needs to be called before running this function
}
\examples{

my_phemdObj <- createDataObj(all_expn_data, all_genes, as.character(snames_data))
my_phemdObj_lg <- removeTinySamples(my_phemdObj, 10)
my_phemdObj_lg <- aggregateSamples(my_phemdObj_lg, max_cells=1000)
my_phemdObj_lg <- selectFeatures(my_phemdObj_lg, selected_genes=c('TP53', 
'EGFR', 'KRAS', 'FOXP3', 'LAG3'))

}
